// src/message/message.enum.ts
export enum MessageType {
    SEND_MESSAGE = 'SEND_MESSAGE',
    RECEIVE_MESSAGE = 'RECEIVE_MESSAGE',
    SEND_UPDATE_MESSAGE = 'SEND_UPDATE_MESSAGE',
    RECEIVE_UPDATED_MESSAGE = 'RECEIVE_UPDATED_MESSAGE',
    SEND_MESSAGE_SEEN = 'SEND_MESSAGE_SEEN',
    RECEIVE_MESSAGE_SEEN = 'RECEIVE_MESSAGE_SEEN',
    SEND_DELETE_MESSAGE = 'SEND_DELETE_MESSAGE',
    RECEIVE_DELETED_MESSAGE = 'RECEIVE_DELETED_MESSAGE', 
}